+++ src/main/java/de/alpharogroup/bean/mapper/GenericMapper.java
@@ -34,7 +34,7 @@ import org.dozer.MappingException;
 import lombok.NonNull;
 
 /**
- * The Interface {@link EntityDOMapper} provides the methods for mapping entities to domain objects
+ * The Interface {@link GenericMapper} provides the methods for mapping entities to domain objects
  * and back.
  *
  * @param <E>
@@ -42,7 +42,7 @@ import lombok.NonNull;
  * @param <DO>
  *            the generic type of the domain object
  */
-public interface EntityDOMapper<E, DO>
+public interface GenericMapper<E, DO>
 {
 
 	/**
@@ -106,38 +106,6 @@ public interface EntityDOMapper<E, DO>
 		throws MappingException
 	{
 		return MapperExtensions.map(getMapper(), source, destinationClass);
-	};
-
-	/**
-	 * Maps the given entity object to a domain object.
-	 *
-	 * @param entity
-	 *            the entity
-	 * @return the domain object
-	 * @deprecated use instead method <code>toDto</code><br>
-	 *             <br>
-	 *             Note: will be removed on next minor version
-	 */
-	@Deprecated
-	default DO toDomainObject(final @NonNull E entity)
-	{
-		return toDto(entity);
-	};
-
-	/**
-	 * Maps the given list of entity objects to a list of domain objects.
-	 *
-	 * @param entities
-	 *            the entities
-	 * @return the list of domain objects.
-	 * @deprecated use instead method <code>toDtos</code><br>
-	 *             <br>
-	 *             Note: will be removed on next minor version
-	 */
-	@Deprecated
-	default List<DO> toDomainObjects(final @NonNull Collection<E> entities)
-	{
-		return toDtos(entities);
 	};
 
 	/**